Correção no modulo Navers

+++ src/modules/Naver/Services/DeleteNaver.ts
@@ -2,6 +2,13 @@ import { injectable, inject } from 'tsyringe';
 
 import INaverRepository from '@modules/Naver/repositories/INaverRepository';
 
+import AppError from '@shared/errors/AppError';
+
+interface Request {
+  id: string;
+  user_id: string;
+}
+
 @injectable()
 export default class DeleteProject {
   constructor(
@@ -9,7 +16,19 @@ export default class DeleteProject {
     private naverRepository: INaverRepository,
   ) {}
 
-  public async execute(id: string): Promise<void> {
+  public async execute({ id, user_id }: Request): Promise<void> {
+    const project = await this.naverRepository.findOneNaver({
+      where: { id },
+    });
+
+    if (!project) {
+      throw new AppError('Nave not found.');
+    }
+
+    if (project.user_id !== user_id) {
+      throw new AppError('This nave does not belong to you.');
+    }
+
     await this.naverRepository.delete(id);
   }
 }
